Upload secret message reveal
Shows the original image the clue.bmp. The secret message is revealed when you remove the shades of red but it's still too light to see very well so the final image is how it looks with the light shades of blue enhanced to make it darker and more visible.